Refactor(backend):Refatoração Entidade Bancada

+++ backend/src/main/java/br/senai/projeto/entity/Bancada.java
@@ -19,7 +19,7 @@ public class Bancada {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_bancada")
-    private Long idBancada;
+    private int idBancada;
 
     @Column(name = "modelo", nullable = false)
     private String modelo;
@@ -33,7 +33,7 @@ public class Bancada {
     private StatusUsoBancada statusUso;
 
     @Column(name = "tempo_base_min")
-    private Integer tempoBaseMin;
+    private int tempoBaseMin;
 
     @ManyToOne
     @JoinColumn(name = "id_unidade", nullable = false)
